rover_py/pyproject.toml
[project]
name = "rover_py"
version = "0.0.1"
dependencies = [
    "python-can>=4.2.2",
    "keyboard>=0.13.5",
    "cantools>=39.3.0",
]
authors = [
  { name="CanEduDev", email="support@canedudev.com" },
]
description = "Helpers for interacting with the CanEduDev Rover"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/canedudev/rover"
Issues = "https://github.com/canedudev/rover/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rover"]
